pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "klimbook-release"
version = "0.1.0"
description = "CLI tool for automated multi-platform release notes generation for Klimbook"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "zxxz6"},
]
dependencies = [
    "anthropic>=0.40.0",
    "gitpython>=3.1.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
klimbook-release = "klimbook_release.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
klimbook_release = ["default_config.yaml"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
